requirements.txt
fastapi
uvicorn
openai
weaviate-client
tiktoken
dotenv